# docs/swagger.yaml
basePath: /
definitions:
...
  app.ListRes:
    properties:
      list:
        description: Data list // 数据清单
      pager:
        allOf:
        - $ref: '#/definitions/app.Pager'
        description: Pagination info // 翻页信息
    type: object
  app.Pager:
    properties:
      page:
        description: Page number // 页码
        type: integer
      pageSize:
        description: Page size // 每页数量
        type: integer
      totalRows:
        description: Total rows // 总行数
        type: integer
    type: object
  app.Res:
    properties:
      code:
        type: integer
      context: {}
      data: {}
      details: {}
      message: {}
      pageIndex:
        description: '1-BASED wire value: 0/absent = non-paginated message, n>0 =
          download page n-1 (internal 0-based). 1-based so page 0 stays distinguishable
          from non-paginated messages under proto3 zero-value elision; only set for
          paginated download messages on pv>=2 connections. Client-to-server PageAck.pageIndex
          stays 0-based. // 线上值为 1-BASED：0/缺省 = 非分页消息，n>0 = 下载页 n-1（内部 0-based）。采用
          1-based 是为了让第 0 页在 proto3 零值不编码规则下仍可与非分页消息区分；仅 pv>=2 连接的分页下载消息会设置。客户端→服务端的
          PageAck.pageIndex 保持 0-based'
      path: {}
      status:
        type: boolean
      vault: {}
    type: object
...
  dto.SourceProbeDTO:
    description: Probe result containing GitHub/CNB reachability, latency, recommended
      source and current mode
    properties:
      cnb:
        allOf:
        - $ref: '#/definitions/dto.SourceProbeItem'
        description: CNB probe result // CNB 探测结果
      github:
        allOf:
        - $ref: '#/definitions/dto.SourceProbeItem'
        description: GitHub probe result // GitHub 探测结果
      recommended:
        description: 'Recommended source: "github" or "cnb" // 推荐源'
        example: github
        type: string
      selectedMode:
        description: Current configured pull-source mode // 当前配置的选源模式
        example: auto
        type: string
    type: object
  dto.SourceProbeItem:
    description: 'Single source probe result: reachability and latency'
    properties:
      latencyMs:
        description: Round-trip latency in ms // 往返延迟（毫秒）
        example: 280
        type: integer
      ok:
        description: Whether the source is reachable // 该源是否可达
        example: true
        type: boolean
    type: object
...
host: localhost:9000
info:
  contact:
    email: haierkeys@gmail.com
    name: Haierkeys
    url: https://github.com/ZC-eto
  description: This is the Fast Note Sync Service HTTP API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Fast Note Sync Service HTTP API
  version: "1.0"
paths:
...
  /api/support:
    get:
      description: Get support records for the specified language with pagination
        and sorting
      parameters:
      - description: 'Language code (default: en)'
        in: query
        name: lang
        type: string
      - description: Sort by field (amount, time, name, item)
        in: query
        name: sortBy
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sortOrder
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/app.ListRes'
              type: object
      summary: Get support records
      tags:
      - System
...
  /api/version/probe:
    get:
      description: Parallel-probe GitHub and CNB release endpoints, return reachability,
        latency, recommended source and current selected mode
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/app.Res'
            - properties:
                data:
                  $ref: '#/definitions/dto.SourceProbeDTO'
              type: object
      security:
      - UserAuthToken: []
      summary: Probe version sources latency
      tags:
      - System
...
securityDefinitions:
  ShareAuthToken:
    in: header
    name: Share-Token
    type: apiKey
  UserAuthToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"